login-und-register refactored to include no logic
