+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/ui/LoginScreen.kt
@@ -1,22 +1,28 @@
 package at.aau.serg.websocketbrokerdemo.ui
 
-import android.content.Context
-import android.content.Intent
-import android.app.Activity
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.material3.Button
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.navigation.NavController
+import at.aau.serg.websocketbrokerdemo.auth.AuthManager
 import com.google.firebase.auth.FirebaseAuth
-import at.aau.serg.websocketbrokerdemo.MainActivity
-import com.google.firebase.FirebaseException
-import com.google.firebase.auth.FirebaseAuthException
+
 
 @Composable
 fun LoginScreen(navController: NavController) {
@@ -54,7 +60,7 @@ fun LoginScreen(navController: NavController) {
             Text(it, color = MaterialTheme.colorScheme.error)
         }
 
-        Button(onClick = { loginUser(auth, email, password, context, setErrorMessage) }) {
+        Button(onClick = { AuthManager.loginUser(auth, email, password, context, setErrorMessage) }) {
             Text("Login")
         }
 
@@ -66,31 +72,4 @@ fun LoginScreen(navController: NavController) {
 
 
 
-private fun loginUser(auth: FirebaseAuth, email: String, password: String, context: Context, setErrorMessage: (String) -> Unit) {
-    auth.signInWithEmailAndPassword(email, password)
-        .addOnSuccessListener {
-            context.startActivity(Intent(context, MainActivity::class.java))
-            (context as? Activity)?.finish()
-        }
-        .addOnFailureListener { error ->
-            (context as? Activity)?.runOnUiThread {
-                val errorMessage = when (error) {
-                    is FirebaseAuthException -> {
-                        when (error.errorCode) {
-                            "ERROR_WRONG_PASSWORD" -> "The password is incorrect. Please try again."
-                            "ERROR_INVALID_EMAIL" -> "The email address is invalid. Please check and try again."
-                            "ERROR_USER_NOT_FOUND" -> "No account found with this email. Please register first."
-                            else -> "Login failed: ${error.message ?: "Unknown error."}"
-                        }
-                    }
-                    is FirebaseException -> {
-                        "An error occurred. Please check your credentials."
-                    }
-                    else -> {
-                        "Unknown error: ${error.message ?: "Unknown error."}"
-                    }
-                }
-                setErrorMessage(errorMessage)
-            }
-        }
-}
+
